feat(openai): split mod.rs in to two
